requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
Pillow==9.2.0
Werkzeug==3.0.4
